Refactor reading history handling and modal
Move reading history logic to a dedicated handler (src/utils/readingHistoryHandler.js) and remove the old readingHistory.js. Introduce a __readingHistoryModalVisible store and wire it into the ReadingHistoryModal component and UI (layout and homepage) so the modal is controlled via store rather than local state. Add auto-close behavior for the modal on navigation and reset the active reading session when leaving reader pages. Update updateSettings to update the lastRead store and then sync readingHistory using the new handler. Also include minor UI tweaks (ReadingHistoryModal text/layout and button tooltip) and a simplified Clock SVG path.
